Rust 分离菜单与托盘，新增语言配置
