+++ src-tauri/Cargo.toml
@@ -14,8 +14,8 @@ rust-version = "1.57"
 tauri-build = { version = "1.0.0", features = [] }
 
 [dependencies]
-serde_json = "1.0"
-serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.85"
+serde = { version = "1.0.144", features = ["derive"] }
 tauri = { version = "1.0.5", features = ["api-all", "icon-png", "macos-private-api", "system-tray"] }
 window-vibrancy = "0.1.3"
