+++ src-tauri/json/en.json
@@ -1,0 +1,6 @@
+{
+  "tray_show": "Show",
+  "tray_hide": "Show/Hide",
+  "tray_quit": "Quit",
+  "change_ico": "Change Icon"
+}
