+++ src-tauri/json/zh_CN.json
@@ -1,0 +1,6 @@
+{
+  "tray_show": "显示",
+  "tray_hide": "显示/隐藏",
+  "tray_quit": "退出",
+  "change_ico": "图标"
+}
